feat(products): add security for superuser pages

+++ products/views.py
@@ -1,5 +1,7 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+
 from django.db.models import Q
 from django.db.models.functions import Lower
 from .models import Product
@@ -54,9 +56,13 @@ def product_detail(request, product_id):
     }
     return render(request, 'products/product_detail.html', context)
 
-
+@login_required
 def add_product(request):
     """ Superuser can add a product """
+    if no request.user.is_superuser:
+        messages.error(request, 'No access for non-admin users')
+        return redirect(reverse('home'))
+
     if request.method == "POST":
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -74,9 +80,14 @@ def add_product(request):
     }
     return render(request, template, context)
 
+@login_required
 def edit_product(request, product_id):
     """ Superuser can edit a product """
     product = get_object_or_404(Product, pk=product_id)
+
+    if no request.user.is_superuser:
+        messages.error(request, 'No access for non-admin users')
+        return redirect(reverse('home'))
 
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -98,9 +109,15 @@ def edit_product(request, product_id):
 
     return render(request, template, context)
 
+@login_required
 def delete_product(request, product_id):
     """ Superuser can delete product """
     product = get_object_or_404(Product, pk=product_id)
+
+    if no request.user.is_superuser:
+        messages.error(request, 'No access for non-admin users')
+        return redirect(reverse('home'))
+
     product.delete()
     messages.success(request, 'Product deleted')
     return redirect(reverse('products'))
